Rename "index.ts" to "app-state.ts"
