+++ src/app/app.module.ts
@@ -12,8 +12,8 @@ import { AppComponent } from "@app/app.component";
 import { MainModule } from "@app/components/main/main.module";
 import { SOLVER_PROVIDERS } from "@app/core/solver/sudoku-solver.provider";
 import { TranslateHttpLoaderFactory } from "@app/core/translate/translate-loader-factory";
-import { metaReducers, reducers } from "@app/state";
 import { appInitializer } from "@app/state/app-initialization.service";
+import { metaReducers, reducers } from "@app/state/app-state";
 import { EffectsModule } from "@ngrx/effects";
 import { StoreModule, StoreRootModule } from "@ngrx/store";
 import { StoreDevtoolsModule } from "@ngrx/store-devtools";
