+++ src/app/components/sudoku-solver/state/sudoku-solver.selectors.spec.ts
@@ -8,7 +8,7 @@ import {
   SudokuGridViewModel,
   SudokuGridViewModelBranchInfo,
 } from "@app/shared/types/sudoku-grid-view-model";
-import { AppState } from "@app/state";
+import { AppState } from "@app/state/app-state";
 import { Puzzle4x4 } from "@app/test/puzzles/puzzle-4x4";
 import { TestState } from "@app/test/state/test-state";
